Refactor api service
