+++ MovieApp/MovieApp/API/APIClient.swift
@@ -1,0 +1,35 @@
+//
+//  APIClient.swift
+//  MovieApp
+//
+//  Created by Elliot on 3/25/20.
+//  Copyright © 2020 Elliot. All rights reserved.
+//
+
+import RxSwift
+
+class APIClient {
+    
+    private let baseURL = URL(string: AppConstans.baseURL)!
+    
+    func send<T: Codable>(apiRequest: APIRequest) -> Observable<T> {
+        return Observable<T>.create { observer in
+            let request = apiRequest.request(with: URL(string: AppConstans.baseURL)!)
+            let task = URLSession.shared.dataTask(with: request) { (data, response, error) in
+                do {
+                    let model: T = try JSONDecoder().decode(T.self, from: data ?? Data())
+                    observer.onNext(model)
+                } catch let error {
+                    observer.onError(error)
+                }
+                observer.onCompleted()
+            }
+            task.resume()
+            
+            return Disposables.create {
+                task.cancel()
+            }
+        }
+    }
+    
+}
