+++ MovieApp/MovieApp/MoviesList/Services/MovieService.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2020 Elliot. All rights reserved.
 //
 
-import Foundation
 import RxSwift
 
 protocol MovieServiceProtocol {
@@ -14,28 +13,10 @@ protocol MovieServiceProtocol {
 }
 
 class MovieService: MovieServiceProtocol {
-    
-    private let baseURL = URL(string: AppConstans.baseURL)!
-    
+
     func fetchMovies() -> Observable<Result> {
-        return Observable.create {[unowned self] observer in
-            let apiRequest = MovieRequest()
-            let request = apiRequest.request(with: self.baseURL)
-            let task = URLSession.shared.dataTask(with: request) { (data, response, error) in
-                do {
-                    let model = try JSONDecoder().decode(Result.self, from: data ?? Data())
-                    observer.onNext(model)
-                } catch let error {
-                    observer.onError(error)
-                }
-                observer.onCompleted()
-            }
-            task.resume()
-            
-            return Disposables.create {
-                task.cancel()
-            }
-        }
+        APIClient().send(apiRequest: MovieRequest())
     }
-    
 }
+
+
